EDU-211 Move constants.js to domain folder
