+++ src/components/pages/users.js
@@ -3,10 +3,10 @@ import firstBy from 'thenby';
 import autoBind from 'auto-bind';
 import PropTypes from 'prop-types';
 import Logger from '../../common/logger.js';
-import { ROLE } from '../../domain/role.js';
 import { Table, Popover, Menu } from 'antd';
 import { withUser } from '../user-context.js';
 import { withTranslation } from 'react-i18next';
+import { ROLE } from '../../domain/constants.js';
 import { inject } from '../container-context.js';
 import errorHelper from '../../ui/error-helper.js';
 import { withPageName } from '../page-name-context.js';
