+++ src/domain/permissions.spec.js
@@ -1,4 +1,4 @@
-import { ROLE } from './role.js';
+import { ROLE } from './constants.js';
 import { exportUser } from './built-in-users.js';
 import permissions, { hasUserPermission } from './permissions.js';
