--- src/domain/role.js
@@ -1,4 +1,0 @@
-export const ROLE = {
-  user: 'user',
-  admin: 'admin'
-};
