+++ src/server/my-space-controller.js
@@ -2,7 +2,7 @@ import PageRenderer from './page-renderer.js';
 import { PAGE_NAME } from '../domain/page-name.js';
 import RoomService from '../services/room-service.js';
 import ServerConfig from '../bootstrap/server-config.js';
-import { FEATURE_TOGGLES } from '../common/constants.js';
+import { FEATURE_TOGGLES } from '../domain/constants.js';
 import ClientDataMapper from '../server/client-data-mapper.js';
 import needsAuthentication from '../domain/needs-authentication-middleware.js';
