+++ src/services/document-import-task-processor.js
@@ -7,7 +7,7 @@ import UserService from './user-service.js';
 import DocumentService from './document-service.js';
 import ExportApiClient from './export-api-client.js';
 import ServerConfig from '../bootstrap/server-config.js';
-import { DOCUMENT_ORIGIN } from '../common/constants.js';
+import { DOCUMENT_ORIGIN } from '../domain/constants.js';
 
 const logger = new Logger(import.meta.url);
