+++ src/services/import-service.spec.js
@@ -6,7 +6,7 @@ import ImportService from './import-service.js';
 import ExportApiClient from './export-api-client.js';
 import DocumentStore from '../stores/document-store.js';
 import BatchLockStore from '../stores/batch-lock-store.js';
-import { BATCH_TYPE, DOCUMENT_ORIGIN, TASK_TYPE } from '../common/constants.js';
+import { BATCH_TYPE, DOCUMENT_ORIGIN, TASK_TYPE } from '../domain/constants.js';
 import { createTestDocument, destroyTestEnvironment, pruneTestEnvironment, setupTestEnvironment, setupTestUser } from '../test-helper.js';
 
 const { BadRequest } = httpErrors;
